refactor(cypher): move reentry helpers out of lowering
